+++ extensions/dear-imnodes-sys/src/bindings_pregenerated.rs
@@ -209,7 +209,7 @@ pub type ImGuiTreeNodeFlags = ::std::os::raw::c_int;
 pub type ImGuiViewportFlags = ::std::os::raw::c_int;
 pub type ImGuiWindowFlags = ::std::os::raw::c_int;
 pub type ImWchar16 = ::std::os::raw::c_ushort;
-pub type ImWchar = ImWchar16;
+pub type ImWchar = ImWchar32;
 pub type ImGuiSelectionUserData = ImS64;
 pub type ImGuiSizeCallback =
     ::std::option::Option<unsafe extern "C" fn(data: *mut ImGuiSizeCallbackData)>;